Create Triangleproblem.c
Problem
You are given n triangles, specifically, their sides a, b and c. Print them in the same style but sorted by their areas from the smallest one to the largest one. It is guaranteed that all the areas are different.


The best way to calculate a volume of the triangle with sides a, b and c is Heron's formula:

S = [p x (p - a) x (p - b) x (p - c)]^1/2 where p = a + b + c / 2.

Input Format 
First line of each test file contains a single integer n. n lines follow with a, b and c on each separated by single spaces.

Constraints
1 <= n <= 100
1 <= a, b, c <= 70
a + b > c,a + c > b + c > a
Output Format
Print exactly n lines. On each line print 3 integers separated by single spaces, which are a, b and c of the corresponding triangle.

Sample Input 0

3
7 24 25
5 12 13
3 4 5
Sample Output 0

3 4 5
5 12 13
7 24 25
Explanation 0

The square of the first triangle is 84. The square of the second triangle is 30. The square of the third triangle is 6. So the sorted order is the reverse one.