created sweep.ino
Sweeps the shaft of a RC servo motor back and forth across 180 degrees.

Servo motors have three wires: power, ground, and signal. The power wire is typically red, and should be connected to the 5V pin on the Arduino board. The ground wire is typically black or brown and should be connected to a ground pin on the board. The signal pin is typically yellow, orange or white and should be connected to pin 9 on the board.

For detail please visit:
https://www.arduino.cc/en/Tutorial/LibraryExamples/Sweep